feat: updated copy for seo
