+++ src/app/posts/[slug]/page.tsx
@@ -108,7 +108,7 @@ const SinglePostPage = () => {
               Resources
             </Link>
             <span className="mx-3 text-gray-300">›</span>
-            <Link href="//toasty-tidbits-archives" className="hover:text-gray-700 transition-colors">
+            <Link href="/toasty-tidbits-archives" className="hover:text-gray-700 transition-colors">
               Archives
             </Link>
             <span className="mx-3 text-gray-300">›</span>
@@ -247,7 +247,7 @@ const SinglePostPage = () => {
               </div>
 
               <div className="mt-12 text-center">
-                <Link href="//toasty-tidbits-archives">
+                <Link href="/toasty-tidbits-archives">
                   <Button className="bg-tst-yellow">View all stories</Button>
                 </Link>
               </div>
