+++ src/data/resourceData.ts
@@ -31,6 +31,6 @@ export const resourcesPageData = {
   },
   routes: {
     privacyPolicy: "/policy",
-    newsletterArchives: "//toasty-tidbits-archives"
+    newsletterArchives: "/toasty-tidbits-archives"
   }
 };
